LayouterApllicator tests added; fixes;

+++ TagsCloudContainer/WordInfo.cs
@@ -6,7 +6,7 @@ namespace TagsCloudContainer
     {
         public string Word { get; set; }
         public Rectangle Rect { get; set; }
-        public int FontSize { get; set; }
+        public float FontSize { get; set; }
         public int Frequency { get; set; }
     }
 }
